+++ crates/robin_engine/src/engine/level_loading.rs
@@ -3060,7 +3060,7 @@ impl EngineInner {
                     let ransom = self
                         .campaign
                         .as_ref()
-                        .map(|c| c.get_value(crate::campaign::CampaignValue::Ransom as usize))
+                        .map(|c| c.get_value(crate::campaign::CampaignValue::Ransom))
                         .unwrap_or(0);
                     let purse_threshold = crate::inventory::COINS_PER_PURSE as i32
                         * crate::inventory::COIN_VALUE as i32;
